Add cache configuration, media discovery, and processing modules
- Introduced `cache_config.py` to manage cache settings and a global cache instance.
- Enhanced `config.py` to support YAML configuration loading and allow duplicate keys.
- Created `media_discovery.py` for discovering media IDs from Plex API and folder scanning.
- Developed `media_processing.py` to handle individual media item processing, including scraping and data comparison.
- Implemented `orchestrator.py` to coordinate the main execution flow of the scraper.
- Added `scheduler.py` for scheduling script execution using cron expressions.
- Refactored `scraper.py` to include a new function for initializing the WebDriver and logging in.
- Improved error handling and logging throughout the modules.
